feat: Add ability to configure if marhsal and/or unmarshal method should be generated when using TSMessageAttribute + generated TS messages are now in the same namespace as where they are declared in C#
