+++ src/SourceGenerators/Uno.UI.SourceGenerators/Uno.UI.SourceGenerators.csproj
@@ -95,6 +95,9 @@
 		<Compile Include="..\..\Uno.UI\UI\Xaml\XamlFilePathHelper.shared.cs">
 			<Link>Helpers\XamlFilePathHelper.shared.cs</Link>
 		</Compile>
+		<Compile Include="..\..\Uno.Foundation.Runtime.WebAssembly\Interop\TSInteropMessageAttribute.wasm.cs">
+			<Link>TSBindings\TSInteropMessageAttribute.cs</Link>
+		</Compile>
 	</ItemGroup>
 
 	<Import Project="..\..\Uno.Foundation\Uno.Core.Extensions\Uno.Core.Extensions.props" />
