+++ src/Uno.Foundation.Runtime.WebAssembly/Interop/TSInteropMessageAttribute.wasm.cs
@@ -9,6 +9,16 @@ namespace Uno.Foundation.Interop
 	/// </summary>
 	[AttributeUsage(AttributeTargets.Struct, AllowMultiple = false)]
 	public class TSInteropMessageAttribute : Attribute
+	{	
+		public CodeGeneration Marshaller { get; set; }
+
+		public CodeGeneration UnMarshaller { get; set; }
+	}
+
+	public enum CodeGeneration
 	{
+		Auto = 0,
+		Enabled = 1,
+		Disabled = 256,
 	}
 }
