+++ src/Uno.UI.Wasm.Tests/WasmScripts/UnitTests.d.ts
@@ -12,30 +12,35 @@ declare class TSBindingsTests {
     When_NullArrayOfStrings(pParams: number, pReturn: number): boolean;
     When_ArrayOfNullStrings(pParams: number, pReturn: number): boolean;
 }
-declare class GenericReturn {
-    Value: string;
-    marshal(pData: number): void;
+declare namespace SamplesApp.UnitTests.TSBindings {
+    class GenericReturn {
+        Value: string;
+        marshal(pData: number): void;
+    }
 }
-declare class When_ArrayOfIntParams {
-    MyArray_Length: number;
-    MyArray: Array<number>;
-    static unmarshal(pData: number): When_ArrayOfIntParams;
+declare namespace SamplesApp.UnitTests.TSBindings {
+    class When_ArrayOfIntParams {
+        MyArray_Length: number;
+        MyArray: Array<number>;
+        static unmarshal(pData: number): When_ArrayOfIntParams;
+    }
 }
-declare class When_ArrayOfStringParams {
-    MyArray_Length: number;
-    MyArray: Array<string>;
-    static unmarshal(pData: number): When_ArrayOfStringParams;
+declare namespace SamplesApp.UnitTests.TSBindings {
+    class When_ArrayOfStringsParams {
+        MyArray_Length: number;
+        MyArray: Array<string>;
+        static unmarshal(pData: number): When_ArrayOfStringsParams;
+    }
 }
-declare class When_ArrayOfStringsParams {
-    MyArray_Length: number;
-    MyArray: Array<string>;
-    static unmarshal(pData: number): When_ArrayOfStringsParams;
+declare namespace SamplesApp.UnitTests.TSBindings {
+    class When_IntPtrParams {
+        Id: number;
+        static unmarshal(pData: number): When_IntPtrParams;
+    }
 }
-declare class When_IntPtrParams {
-    Id: number;
-    static unmarshal(pData: number): When_IntPtrParams;
-}
-declare class When_SingleStringParams {
-    MyString: string;
-    static unmarshal(pData: number): When_SingleStringParams;
+declare namespace SamplesApp.UnitTests.TSBindings {
+    class When_SingleStringParams {
+        MyString: string;
+        static unmarshal(pData: number): When_SingleStringParams;
+    }
 }
