+++ src/Uno.UI.Wasm.Tests/ts/TSBindingsTests.ts
@@ -5,16 +5,15 @@
 });
 
 class TSBindingsTests {
-
 	public TSBindingsTests() {
 
 	}
 
 	public When_IntPtr(pParams: number, pReturn: number): boolean {
 
-		var params = When_IntPtrParams.unmarshal(pParams);
+		var params = SamplesApp.UnitTests.TSBindings.When_IntPtrParams.unmarshal(pParams);
 
-		var ret = new GenericReturn();
+		var ret = new SamplesApp.UnitTests.TSBindings.GenericReturn();
 		ret.Value = String(params.Id);
 
 		ret.marshal(pReturn);
@@ -24,9 +23,9 @@ class TSBindingsTests {
 
 	public When_IntPtr_Zero(pParams: number, pReturn: number): boolean {
 
-		var params = When_IntPtrParams.unmarshal(pParams);
+		var params = SamplesApp.UnitTests.TSBindings.When_IntPtrParams.unmarshal(pParams);
 
-		var ret = new GenericReturn();
+		var ret = new SamplesApp.UnitTests.TSBindings.GenericReturn();
 		ret.Value = String(params.Id);
 
 		ret.marshal(pReturn);
@@ -36,9 +35,9 @@ class TSBindingsTests {
 
 	public When_SingleString(pParams: number, pReturn: number): boolean {
 
-		var params = When_SingleStringParams.unmarshal(pParams);
+		var params = SamplesApp.UnitTests.TSBindings.When_SingleStringParams.unmarshal(pParams);
 
-		var ret = new GenericReturn();
+		var ret = new SamplesApp.UnitTests.TSBindings.GenericReturn();
 		ret.Value = params.MyString;
 
 		ret.marshal(pReturn);
@@ -48,9 +47,9 @@ class TSBindingsTests {
 
 	public When_SingleUnicodeString(pParams: number, pReturn: number): boolean {
 
-		var params = When_SingleStringParams.unmarshal(pParams);
+		var params = SamplesApp.UnitTests.TSBindings.When_SingleStringParams.unmarshal(pParams);
 
-		var ret = new GenericReturn();
+		var ret = new SamplesApp.UnitTests.TSBindings.GenericReturn();
 		ret.Value = params.MyString;
 
 		ret.marshal(pReturn);
@@ -60,9 +59,9 @@ class TSBindingsTests {
 
 	public When_NullString(pParams: number, pReturn: number): boolean {
 
-		var params = When_SingleStringParams.unmarshal(pParams);
+		var params = SamplesApp.UnitTests.TSBindings.When_SingleStringParams.unmarshal(pParams);
 
-		var ret = new GenericReturn();
+		var ret = new SamplesApp.UnitTests.TSBindings.GenericReturn();
 		ret.Value = String(params.MyString === null);
 
 		ret.marshal(pReturn);
@@ -72,9 +71,9 @@ class TSBindingsTests {
 
 	public When_ArrayOfInt(pParams: number, pReturn: number): boolean {
 
-		var params = When_ArrayOfIntParams.unmarshal(pParams);
+		var params = SamplesApp.UnitTests.TSBindings.When_ArrayOfIntParams.unmarshal(pParams);
 
-		var ret = new GenericReturn();
+		var ret = new SamplesApp.UnitTests.TSBindings.GenericReturn();
 		ret.Value = params.MyArray.join(";");
 
 		ret.marshal(pReturn);
@@ -84,9 +83,9 @@ class TSBindingsTests {
 
 	public When_NullArrayOfInt(pParams: number, pReturn: number): boolean {
 
-		var params = When_ArrayOfIntParams.unmarshal(pParams);
+		var params = SamplesApp.UnitTests.TSBindings.When_ArrayOfIntParams.unmarshal(pParams);
 
-		var ret = new GenericReturn();
+		var ret = new SamplesApp.UnitTests.TSBindings.GenericReturn();
 		ret.Value = String(params.MyArray === null);
 
 		ret.marshal(pReturn);
@@ -96,9 +95,9 @@ class TSBindingsTests {
 
 	public When_ArrayOfStrings(pParams: number, pReturn: number): boolean {
 
-		var params = When_ArrayOfStringsParams.unmarshal(pParams);
+		var params = SamplesApp.UnitTests.TSBindings.When_ArrayOfStringsParams.unmarshal(pParams);
 
-		var ret = new GenericReturn();
+		var ret = new SamplesApp.UnitTests.TSBindings.GenericReturn();
 		ret.Value = params.MyArray.join(";");
 
 		ret.marshal(pReturn);
@@ -108,9 +107,9 @@ class TSBindingsTests {
 
 	public When_ArrayOfUnicodeStrings(pParams: number, pReturn: number): boolean {
 
-		var params = When_ArrayOfStringsParams.unmarshal(pParams);
+		var params = SamplesApp.UnitTests.TSBindings.When_ArrayOfStringsParams.unmarshal(pParams);
 
-		var ret = new GenericReturn();
+		var ret = new SamplesApp.UnitTests.TSBindings.GenericReturn();
 		ret.Value = params.MyArray.join(";");
 
 		ret.marshal(pReturn);
@@ -120,9 +119,9 @@ class TSBindingsTests {
 
 	public When_NullArrayOfStrings(pParams: number, pReturn: number): boolean {
 
-		var params = When_ArrayOfStringsParams.unmarshal(pParams);
+		var params = SamplesApp.UnitTests.TSBindings.When_ArrayOfStringsParams.unmarshal(pParams);
 
-		var ret = new GenericReturn();
+		var ret = new SamplesApp.UnitTests.TSBindings.GenericReturn();
 		ret.Value = String(params.MyArray === null);
 
 		ret.marshal(pReturn);
@@ -132,9 +131,9 @@ class TSBindingsTests {
 
 	public When_ArrayOfNullStrings(pParams: number, pReturn: number): boolean {
 
-		var params = When_ArrayOfStringsParams.unmarshal(pParams);
+		var params = SamplesApp.UnitTests.TSBindings.When_ArrayOfStringsParams.unmarshal(pParams);
 
-		var ret = new GenericReturn();
+		var ret = new SamplesApp.UnitTests.TSBindings.GenericReturn();
 		ret.Value = params.MyArray.map(v => v === null).join(";");
 
 		ret.marshal(pReturn);
