+++ src/Uno.UI.Wasm.Tests/tsbindings/SamplesApp_UnitTests_TSBindings_When_SingleStringParams.ts
@@ -1,0 +1,27 @@
+/* TSBindingsGenerator Generated code -- this code is regenerated on each build */
+namespace SamplesApp.UnitTests.TSBindings
+{
+	export class When_SingleStringParams
+	{
+		/* Pack=1 */
+		public MyString : string;
+		public static unmarshal(pData:number) : When_SingleStringParams
+		{
+			const ret = new When_SingleStringParams();
+			
+			{
+				const ptr = Module.getValue(pData + 0, "*");
+				if(ptr !== 0)
+				{
+					ret.MyString = String(Module.UTF8ToString(ptr));
+				}
+				else
+				
+				{
+					ret.MyString = null;
+				}
+			}
+			return ret;
+		}
+	}
+}
