--- src/Uno.UI/tsBindings/ApplicationDataContainer_ContainsKeyReturn.ts
@@ -1,10 +1,0 @@
-/* TSBindingsGenerator Generated code -- this code is regenerated on each build */
-class ApplicationDataContainer_ContainsKeyReturn
-{
-	/* Pack=4 */
-	public ContainsKey : boolean;
-	public marshal(pData:number)
-	{
-		Module.setValue(pData + 0, this.ContainsKey, "i32");
-	}
-}
