--- src/Uno.UI/tsBindings/ApplicationDataContainer_GetCountReturn.ts
@@ -1,10 +1,0 @@
-/* TSBindingsGenerator Generated code -- this code is regenerated on each build */
-class ApplicationDataContainer_GetCountReturn
-{
-	/* Pack=4 */
-	public Count : number;
-	public marshal(pData:number)
-	{
-		Module.setValue(pData + 0, this.Count, "i32");
-	}
-}
