--- src/Uno.UI/tsBindings/ApplicationDataContainer_RemoveReturn.ts
@@ -1,10 +1,0 @@
-/* TSBindingsGenerator Generated code -- this code is regenerated on each build */
-class ApplicationDataContainer_RemoveReturn
-{
-	/* Pack=1 */
-	public Removed : boolean;
-	public marshal(pData:number)
-	{
-		Module.setValue(pData + 0, this.Removed, "i32");
-	}
-}
